+++ example/traffic.ts
@@ -1,0 +1,77 @@
+import 'reflect-metadata'
+import { StateMachine, IStateDeclaration } from '../src/index'
+
+enum Colours {
+  Red = 'Red',
+  Orange = 'Orange',
+  Green = 'Green'
+}
+
+type TrafficLightProperties = IStateDeclaration<TrafficLightStateMachine>;
+type TrafficLightStates = Colours.Red|Colours.Orange|Colours.Green
+
+/**
+ * Defines a state machine to simulate a Traffic Light.
+ *
+ * Can be in Red, Orange, or Green state. Each state has a unique "message"
+ * associated with it.
+ */
+class TrafficLightStateMachine extends StateMachine<TrafficLightProperties, TrafficLightStates> {
+  message: string
+  safe: boolean
+
+  @StateMachine.extend(StateMachine.INITIAL, [Colours.Green])
+  [Colours.Red]: TrafficLightProperties = {
+    message: 'STOP'
+  }
+
+  @StateMachine.extend(StateMachine.INITIAL, [Colours.Green, Colours.Red])
+  [Colours.Orange]: TrafficLightProperties = {
+    message: 'CAUTION'
+  }
+
+  @StateMachine.extend(StateMachine.INITIAL, Colours.Orange)
+  [Colours.Green]: TrafficLightProperties = {
+    message: 'GO',
+    safe: true
+  }
+
+  constructor () {
+    super({
+      // Tells the state machine what state(s) it can initially transition to
+      initialTransitions: [Colours.Green],
+
+      // Define the initial state values for the machine
+      initialStateProperties: {
+        message: 'OFF',
+        safe: false
+      }
+    })
+  }
+}
+
+const machine = new TrafficLightStateMachine()
+const invalidStateName = 'Blue'
+
+const logMachineState = (c: string) => {
+  console.log('========================================')
+  console.log(`Entered ${c} state.`)
+  console.log(`Light is ${machine.currentState}. Message is "${machine.message}". Safe: ${machine.safe}`)
+  console.log('========================================\n')
+}
+
+Object.values(Colours).forEach(c => {
+  machine.onEnter(c, () => logMachineState(c))
+})
+logMachineState(StateMachine.INITIAL)
+machine.transitTo(Colours.Green)
+machine.transitTo(Colours.Orange)
+machine.transitTo(Colours.Red)
+
+// This transition will fail, even if we bypass type safety
+const failReason = machine.transitTo(invalidStateName as any)
+
+if (failReason) {
+  console.log(`Transition to ${invalidStateName} failed. Reason: ${failReason}`)
+}
+