+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# 2.0.0
+
+* `StateMachine` constructor now expects an options Object, i.e `super(opts)`.
+* `opts.initialTransitions` is required by the super constructor. This replaces `$next` requirement.
+* `opts.initialStateProperties` is required by the super constructor. This replaces `rememberInitState` requirement.
+* Add `Properties` and `ValidStates` generics, i.e `StateMachine<Properties, ValidStates>`
+* Disable `console.error` logs by default. Can be enabled via `opts.logging`.
+* Make `transitTo` and others type safe using `ValidStates` generic.
+* Return typed errors from `transitTo`. These use the `TransitionError` type.
+
 # 1.1.6
 
 * Ensure updated JS files are published using `prepublish` hook.